requirements.txt
fastpip == 1.7.0
fastapi == 0.103.2
uvicorn == 0.23.2
jupyter == 1.0.0
jupyterlab == 4.0.6
matplotlib == 3.7.1
japanize_matplotlib == 1.1.3
regex == 2023.10.3
pydantic == 2.4.2
fastapi_csrf_protect == 0.2.1
motor == 2.5.1
PyJWT == 2.3.0
passlib == 1.7.4
python-decouple == 3.3
dnspython == 2.1.0
gunicorn == 20.1.0
langchain == 0.0.311
python-dotenv == 1.0.0
